backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.30.6
sqlalchemy==2.0.41
psycopg[binary]==3.2.10
alembic==1.16.4
pydantic[email]==2.11.7
python-jose==3.5.0
bcrypt==4.3.0
python-multipart==0.0.20
passlib==1.7.4
pyotp==2.9.0
qrcode==8.0
Pillow==10.4.0
fastapi-mail==1.4.1
python-dateutil
tzdata
openai>=1.40.0
python-dotenv==1.0.1
